fix: only create finance zone if financing is enabled
* Update main.lua

There is no need to create a zone if the player has not enabled the financing option.

* Update main.lua

There is no need to create a zone if the player has not enabled the financing option. In other words, after this update, the zone will not be created if you have not enabled the financing option.