refactor!: move vehicles to shared config (#97)
* refactor: move vehicles to shared config

Trusting the client to tell us what vehicles are in what shop is a bad idea. This fixes that and moves the vehicles to the shared config and provides additional sanity checks to prevent abuse from net event injection. Additionally removes zone check logic from client to server as we shouldn't trust the client not to lie to us. Also adds ability to block certain models from being sold at the vehicle shop.

* fix(server): check test drives as well